Simplify key pair handling and storage
