+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/PrivacyParameters.java
@@ -16,8 +16,8 @@ package org.hyperledger.besu.ethereum.core;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
+import org.hyperledger.besu.crypto.KeyPairUtil;
 import org.hyperledger.besu.crypto.SECP256K1;
-import org.hyperledger.besu.crypto.SECP256K1.KeyPair;
 import org.hyperledger.besu.enclave.Enclave;
 import org.hyperledger.besu.enclave.EnclaveFactory;
 import org.hyperledger.besu.ethereum.privacy.storage.PrivacyStorageProvider;
@@ -201,7 +201,7 @@ public class PrivacyParameters {
       return this;
     }
 
-    public PrivacyParameters build() throws IOException {
+    public PrivacyParameters build() {
       final PrivacyParameters config = new PrivacyParameters();
       if (enabled) {
         final WorldStateStorage privateWorldStateStorage =
@@ -221,7 +221,7 @@ public class PrivacyParameters {
         config.setEnclave(enclaveFactory.createVertxEnclave(enclaveUrl));
 
         if (privateKeyPath != null) {
-          config.setSigningKeyPair(KeyPair.load(privateKeyPath.toFile()));
+          config.setSigningKeyPair(KeyPairUtil.load(privateKeyPath.toFile()));
         }
       }
       config.setEnabled(enabled);
